Update content from new dev branch, fix some fmt problems
