+++ lib/collection/src/problems/unindexed_field.rs
@@ -235,7 +235,7 @@ fn infer_index_from_field_condition(field_condition: &FieldCondition) -> Vec<Fie
             Match::Any(match_any) => infer_index_from_any_variants(&match_any.any),
             Match::Except(match_except) => infer_index_from_any_variants(&match_except.except),
             Match::TextAny(_match_text_any) => vec![FieldIndexType::Text],
-            Match::Fuzzy(_match_fuzzy) => vec![FieldIndexType::TextFuzzy],
+            Match::Fuzzy(_match_fuzzy) => vec![FieldIndexType::Text],
         })
     }
     if let Some(range_interface) = range {
@@ -520,11 +520,11 @@ impl<'a> Extractor<'a> {
             ExpressionInternal::StrDist {
                 field,
                 query: _,
-                func: _
+                func: _,
             } => {
                 key = field.clone();
                 required_index = vec![FieldIndexType::Text];
-            },
+            }
         }
 
         if self.needs_index(&key, &required_index) {
@@ -546,7 +546,6 @@ enum FieldIndexType {
     FloatRange,
     Text,
     TextPhrase,
-    TextFuzzy,
     BoolMatch,
     UuidMatch,
     UuidRange,
@@ -597,15 +596,10 @@ fn schema_capabilities(value: &PayloadFieldSchema) -> HashSet<FieldIndexType> {
             PayloadSchemaParams::Float(_) => index_types.insert(FieldIndexType::FloatRange),
             PayloadSchemaParams::Geo(_) => index_types.insert(FieldIndexType::Geo),
             PayloadSchemaParams::Text(TextIndexParams {
-                phrase_matching,
-                fuzzy_matching,
-                ..
+                phrase_matching, ..
             }) => {
                 if phrase_matching.unwrap_or_default() {
                     index_types.insert(FieldIndexType::TextPhrase);
-                }
-                if fuzzy_matching.unwrap_or_default() {
-                    index_types.insert(FieldIndexType::TextFuzzy);
                 }
                 index_types.insert(FieldIndexType::Text)
             }
@@ -630,13 +624,6 @@ impl From<FieldIndexType> for PayloadFieldSchema {
                 PayloadFieldSchema::FieldParams(PayloadSchemaParams::Text(TextIndexParams {
                     r#type: TextIndexType::Text,
                     phrase_matching: Some(true),
-                    ..Default::default()
-                }))
-            }
-            FieldIndexType::TextFuzzy => {
-                PayloadFieldSchema::FieldParams(PayloadSchemaParams::Text(TextIndexParams {
-                    r#type: TextIndexType::Text,
-                    fuzzy_matching: Some(true),
                     ..Default::default()
                 }))
             }
