+++ lib/segment/Cargo.toml
@@ -105,8 +105,8 @@ charabia = { version = "0.9.9", default-features = false, features = [
     "chinese-segmentation",
     "chinese-normalization",
 ] }
-fst = { version = "0.4.7", features = ["levenshtein"]}
 strsim = "0.11.1"
+fst = { version = "0.4.7", features = ["levenshtein"]}
 
 gridstore = { path = "../gridstore" }
